+++ pkg/model/model.go
@@ -41,9 +41,9 @@ type Coordinate struct {
 // Sector represents a 2D arc emanating from a location
 type Sector struct {
 	Center  Coordinate `mapstructure:"center"`
-	Azimuth int32      `mapstructure:"azimuth"`
+	Azimuth float64    `mapstructure:"azimuth"`
 	Arc     int32      `mapstructure:"arc"`
-	Tilt    int32      `mapstructure:"tilt"`
+	Tilt    float64    `mapstructure:"tilt"`
 	Height  int32      `mapstructure:"height"`
 }
 
@@ -132,8 +132,8 @@ type Cell struct {
 }
 
 type Beam struct {
-	H3dBAngle              uint32  `mapstructure:"h3dBAngle"`
-	V3dBAngle              uint32  `mapstructure:"v3dBAngle"`
+	H3dBAngle              float64 `mapstructure:"h3dBAngle"`
+	V3dBAngle              float64 `mapstructure:"v3dBAngle"`
 	MaxGain                float64 `mapstructure:"maxGain"`
 	MaxAttenuationDB       float64 `mapstructure:"maxAttenuationDB"`
 	VSideLobeAttenuationDB float64 `mapstructure:"vSideLobeAttenuationDB"`
